move engine code to separate modules
