+++ frontend/api/bindings.ts
@@ -55,13 +55,13 @@ export function deleteSteamAppinfoCache() {
 }
 
 export type ModLoaderData = { id: string; path: string; mods: Mod[] }
-export type OperatingSystem = "Unknown" | "Linux" | "Windows"
 export type GameEngineBrand = "Unity" | "Unreal" | "Godot"
+export type Game = { id: string; name: string; discriminator: string | null; fullPath: string; architecture: Architecture; scriptingBackend: UnityScriptingBackend; operatingSystem: OperatingSystem; steamLaunch: SteamLaunchOption | null; installedMods: string[]; engine: GameEngine | null; thumbnailUrl: string | null }
 export type UnityScriptingBackend = "Il2Cpp" | "Mono" | "Unknown"
-export type Architecture = "Unknown" | "X64" | "X86"
+export type GameEngine = { brand: GameEngineBrand; version: GameEngineVersion | null }
 export type SteamLaunchOption = { launchId: string; appId: number; description: string | null; executable: string | null; arguments: string | null; appType: string | null; osList: string | null; betaKey: string | null; osArch: string | null }
 export type OwnedGame = { id: string; name: string; installed: boolean; osList: OperatingSystem[]; engine: GameEngineBrand; releaseDate: number; thumbnailUrl: string }
+export type Architecture = "Unknown" | "X64" | "X86"
 export type GameEngineVersion = { major: number; minor: number; patch: number; suffix: string | null; display: string }
-export type GameEngine = { brand: GameEngineBrand; version: GameEngineVersion | null }
+export type OperatingSystem = "Unknown" | "Linux" | "Windows"
 export type Mod = { id: string; name: string; scriptingBackend: UnityScriptingBackend; path: string }
-export type Game = { id: string; name: string; discriminator: string | null; fullPath: string; architecture: Architecture; scriptingBackend: UnityScriptingBackend; operatingSystem: OperatingSystem; steamLaunch: SteamLaunchOption | null; installedMods: string[]; engine: GameEngine | null; thumbnailUrl: string | null }
